requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
deepeval>=0.21.0
playwright>=1.42.0
pytest-playwright>=0.4.4
